Updated to net45 and latest nancy
- Project target framework
- New signatures of ISerializer and IBodyDeserializer
- Nuspec file
- Nuget regerences
- Reset nuspec verison to 0.0.0
