+++ .gitignore
@@ -16,3 +16,4 @@ build/
 [Rr]elease/
 _[Rr]e[Ss]harper.*/
 *.tss
+.vs/
