+++ src/Nancy.Serialization.ServiceStack/Demo/Web.config
@@ -1,18 +1,24 @@
 ﻿<?xml version="1.0"?>
-
 <!--
   For more information on how to configure your ASP.NET application, please visit
   http://go.microsoft.com/fwlink/?LinkId=169433
   -->
-
 <configuration>
+  <!--
+    For a description of web.config changes see http://go.microsoft.com/fwlink/?LinkId=235367.
+
+    The following attributes can be set on the <httpRuntime> tag.
+      <system.Web>
+        <httpRuntime targetFramework="4.5" />
+      </system.Web>
+  -->
   <system.web>
-    <compilation debug="true" targetFramework="4.0" />
+    <compilation debug="true" targetFramework="4.5"/>
     <httpHandlers>
       <add verb="*" type="Nancy.Hosting.Aspnet.NancyHttpRequestHandler" path="*"/>
     </httpHandlers>
+    <pages controlRenderingCompatibilityVersion="4.0"/>
   </system.web>
-
   <system.webServer>
     <modules runAllManagedModulesForAllRequests="true"/>
     <validation validateIntegratedModeConfiguration="false"/>
@@ -20,5 +26,4 @@
       <add name="Nancy" verb="*" type="Nancy.Hosting.Aspnet.NancyHttpRequestHandler" path="*"/>
     </handlers>
   </system.webServer>
-
 </configuration>
