+++ src/Nancy.Serialization.ServiceStack/Nancy.Serialization.ServiceStack/Nancy.Serialization.ServiceStack.nuspec
@@ -2,7 +2,7 @@
 <package xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
   <metadata xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
     <id>Nancy.Serialization.ServiceStack</id>
-    <version>0.6.0</version> <!-- Overridden by rake nuget task but a default is required in nuspec files -->
+    <version>0.0.0</version> <!-- Overridden by rake nuget task but a default is required in nuspec files -->
     <authors>Andreas Håkansson, Steven Robbins and contributors</authors>
     <requireLicenseAcceptance>false</requireLicenseAcceptance>
     <description>Provides JSON (de)serialization support using ServiceStack.Text.</description>
@@ -19,8 +19,8 @@
     <tags>Nancy Json ServiceStack</tags>
   </metadata>
   <files>
-    <file src="build\binaries\Nancy.Serialization.ServiceStack.dll" target="lib\net40" />
-    <file src="build\binaries\Nancy.Serialization.ServiceStack.pdb" target="lib\net40" />
+    <file src="build\binaries\Nancy.Serialization.ServiceStack.dll" target="lib\net45" />
+    <file src="build\binaries\Nancy.Serialization.ServiceStack.pdb" target="lib\net45" />
     <file src="src\Nancy.Serialization.ServiceStack\**\*.cs" target="src" />
   </files>
 </package>
