+++ src/Nancy.Serialization.ServiceStack/Nancy.Serialization.ServiceStack/ServiceStackJsonSerializer.cs
@@ -1,19 +1,20 @@
-﻿namespace Nancy.Serializers.Json.ServiceStack
+﻿namespace Nancy.Serialization.ServiceStack
 {
     using System.Collections.Generic;
     using System.IO;
     using global::ServiceStack.Text;
+    using Responses.Negotiation;
 
     public class ServiceStackJsonSerializer : ISerializer
     {
         /// <summary>
         /// Whether the serializer can serialize the content type
         /// </summary>
-        /// <param name="contentType">Content type to serialise</param>
+        /// <param name="mediaRange">Content type to serialise</param>
         /// <returns>True if supported, false otherwise</returns>
-        public bool CanSerialize(string contentType)
+        public bool CanSerialize(MediaRange mediaRange)
         {
-            return Helpers.IsJsonType(contentType);
+            return Helpers.IsJsonType(mediaRange);
         }
 
         /// <summary>
@@ -28,11 +29,11 @@
         /// <summary>
         /// Serialize the given model with the given contentType
         /// </summary>
-        /// <param name="contentType">Content type to serialize into</param>
+        /// <param name="mediaRange">Content type to serialize into</param>
         /// <param name="model">Model to serialize</param>
         /// <param name="outputStream">Output stream to serialize to</param>
         /// <returns>Serialised object</returns>
-        public void Serialize<TModel>(string contentType, TModel model, Stream outputStream)
+        public void Serialize<TModel>(MediaRange mediaRange, TModel model, Stream outputStream)
         {
             JsonSerializer.SerializeToStream(model, outputStream);
         }
