+++ src/Nancy.Serialization.ServiceStack/Nancy.Serialization.ServiceStack/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="ServiceStack.Text" version="3.9.71" targetFramework="net40" />
+  <package id="ServiceStack.Text" version="3.9.71" targetFramework="net45" />
 </packages>
